Nitpicking about alphabetizing methods, minor OCD thing
