+++ src/main.tsx
@@ -1,44 +1,6 @@
-import { ComponentType, PropsWithChildren } from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
-import {
-  AppState,
-  Auth0Provider,
-  Auth0ProviderOptions,
-  withAuthenticationRequired
-} from '@auth0/auth0-react'
-import { BrowserRouter, useNavigate, Routes, Route } from 'react-router-dom'
-import { Users } from './Test.tsx'
-
-
-const ProtectedUsers = withAuthenticationRequired(Users);
-
-const Auth0ProviderWithRedirectCallback = ({
-  children,
-  ...props
-}: PropsWithChildren<Auth0ProviderOptions>) => {
-  const navigate = useNavigate()
-
-  const onRedirectCallback = (appState?: AppState) => {
-    navigate(appState?.returnTo ?? window.location.pathname)
-  }
-
-  return (
-    <Auth0Provider onRedirectCallback={onRedirectCallback} {...props}>
-      {children}
-    </Auth0Provider>
-  )
-}
-
-const AppRoute = () => {
-  return (
-    <Routes>
-        <Route path="/" element={<App />} />
-        <Route path="/toto" element={<div>toto</div>} />
-        <Route path="/users" element={<ProtectedUsers/>} />
-      </Routes>
-  )
-}
+import { BrowserRouter } from 'react-router-dom'
+import { AppRoute, Auth0ProviderWithRedirectCallback } from './tools/router.tsx'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <BrowserRouter>
